int -> int64 changes for large numbers of obs + datetime fixes

+++ src/CostFunction.h
@@ -28,7 +28,7 @@ public:
 	
 protected:
 	int ls_cnt; 
-	int mObs;
+	int64_t mObs;
 	int nState;
 	real* currState;
 	real* currGradient;
